refactoring transfer data structures
Data is now sent by a channel of pointers (too large to fit in channel).
Multiple changes have been included to ensure a clean running and tested state is maintained.
One consequence of the cleaner transfer is that dispListings, mutes etc are now separated from the listings internal to the sound engine; however, a check has been placed after transfer to warn of a length mismatch between them. This is important as some actions such as delete rely on mutes in particular to maintain parity with internal listings. An out-of-bounds reload would result in an append.

Some optimisations have also been made around the inner loops of the sound engine, naturally there is room for more work but a big improvement has been made. The optimisations were undertaken because of an inital problem with the listings loop (marked "much faster") which caused it to grind to a halt and not produce samples in near enough time to create audio, of an order of many tens of milliseconds. This odd effect was tracked down by profiling, with a new CLI flag to control this.
By following this line of enquiry further Sound Engine listings now have a new type []opSE which has a smaller memory footprint and was the biggest additional improvement in loop time. The unsigned int types may be a bit overzealous (int worked fine) but they will remain until any problems are uncovered. Bounds checks have been placed to mitigate possible overruns.

A few other things that were uncovered in the refactor have been cleaned up along the way, eg. the recover() response has been simplified with the 'stop' channel now returning an int.

Verbose mode has been left disabled to be reintroduced in a separate commit.

Renamed TAPELEN could be changed back to a var in future, for a
consistent tape time at different rates.