refactoring input parsing
This commit is part of an ongoing process to bring the codebase closer in
line to industry standards for better maintainability.
Roadmap for this includes:
- reducing args of parseNewOperation() to systemState only
- eliminating the map check for operators inside parseNewOperation()
- eliminate switch statement to return from parseFunction()
- move var signals to type newListing

The function signature of type processor has been changed to pass and
return a systemState value for efficiency. This saves the runtime having
to do pointer checks when calling a process func.