operator: tape renamed to buff, original functionality replaced by function: tape
The (fka) tape operator previously included a high-pass filter and clipping on input, with
a low-pass filter on output.
However, Synte is a DSP DSL so we can do this in the base language.
This enables the new operator buff to be used more flexibly - for DC
signals, with different filters, tanh instead of clip etc.

The new function 'tape' has been added to be used in place of previous
operator.

The operator tap has also been modified to use the same interpolation as buff, but otherwise remains the same.

This commit does not include all the relevant changes to functions.json
for functions which previously used tape.