removing indiviual listing clipping; replacing with limiter per listing
Previously each listing had a clip on it's output. This was to prevent a listing from ducking the main
output limiter for everything else playing. However, in practice this added quite a bit of distortion
and still ducked the output quite a bit due to the conservative threshold of 4. Even at 1, it still wasn't
ideal.
The kind of situation that can induce this is say, a feedback delay with loop gain > 1.

After some experimentation a limiter per listing was arrived at. This has a 20ms attack and 250ms decay
(if my calculations are correct...) It also does 'over-limiting', aka 'negative-ratio' limiting,
where sounds get quieter the louder they are.
The overall sound is perhaps a little too polite now, this may be revisited at a later date.

The main output limiter is kept as a backstop, as it is has an instant attack and is frequency sensitive.

it may be interesting to experiment with producing pumping effects with the new limiter.