changing verbose mode
Due to a previous change, verbose mode was no longer working.
This is likely due to the pointer was buried in a struct, or for
other reasons no longer pointed to the correct slice and underlying
array.
I don't have time at present to triage this fully, though it would
be interesting to find out and useful for future reference.

However, I have elided the 'code' pointer completely and now
pass by value, which has been tested and works fine.

Now, on each new listing launch the listing display will revert back to
non-verbose mode, which is actually preferable, as in practice verbose
mode is only ever used transiently.