saving debug info in event of sound engine panic
The new function coreDump saves the state of the running listing that
has caused the panic to a new timestamped text file in debug/

save() will report an error if no debug/ dir, and in event of as save
error coreDump will attempt to save to a hidden file instead.
This is not as yet documented, the behaviour may be changed to
add a debug/ dir during a future installation script or in coreDump
itself.