persisting signals in sound engine
previously on reloading a listing, all signals would be reset to
default values.
Signals are now persisted on reload if they have the same name in
both old and new. This makes for a smoother listening experience
and also ensures better continuation of slower signals. The artifical
wait on reload for transient muting has been kept, but reduced.

Persisted signals are matched in transfer() by string comparison.
Although this is not ideal, it is a straightforward approach and an
additional benefit of having the strings inside the sound engine is
that coreDump can use them for a more readable output.

The bool rst has been added to switch off persisting, which is
toggled by the `: reset` command.

If logging is enabled, the new and old state of the listing will be
saved by coreDump(). This may be removed in future.

The embedded struct 'keep' in listingState is no longer necessary so
has been removed. This is because now only the listing and signals
are transferred on reload.
All signals are transferred initially as their position and indexes
may have changed. Persisted signals are then found and re-instated
individually.