adding loudness eq
Due to the competitive nature of digital music production, it has been
observed that many sound systems are calibrated for 'bass down' mixes.
This type of mix reduces the bottom few octaves of audible frequencies,
which is where all the energy is (see README.md section on The Sound
Engine) to allow an equivalent boost in overall gain. This in turn is
compensated by reproduction systems. This is most likely an emergent
effect resulting from the 'loudness wars'.
The loudness eq compensates for this.
At present this is undocumented as a under-the-hood feature that most
won't need to change.