combined commit
Ok, so I messed up. Sorry future self.
After trying to atomically commit a backlog of changes I have given up
because it would be too time consuming to make sure the code is in a
working state for each one. (The last commit wasn't.)
I need to get all this in the repo for backup purposes.
This commit features many changes and tweaks, the most notable being an
overhaul of the limiter(s). The README.md hasn't been fully updated yet