correcting output func
a regression was introduced in recent commits whereby the output would
hang on overload. This was because the running state was not persisted
when stuffing zeros.
The shared variables have been gathered to derive from constants for
better maintenance.
A basic test has been introducted to verify the load threshold.