fixing pause muting, exit envelope
Previously pause relied on the first listing to be unmuted in order to
complete envelope.
The exit envelope has now been dual purposed to serve the same role,
with the benefit that exiting when already paused is faster.
The minimum fadeout has been reduced and dithering moved before the
envelope.