reinstating corrected upsampled interpolation
The problem was found to be the fractional part of x.
Although it is not known at this point whether the original algorithm
was intended to be downsampled, this approach has been found to be
effective in preliminary testing.
During a spreadsheet mock of the upsampling sinc function, which was
found to be accurate, the value 'window' was found to result in further
precision.